+++ src/ActiveForm.php
@@ -70,9 +70,9 @@ class ActiveForm extends \yii\widgets\ActiveForm
      */
     public $fieldClass = 'yii\bootstrap\ActiveField';
     /**
-     * @var array HTML attributes for the form tag. Default is `['role' => 'form']`.
+     * @var array HTML attributes for the form tag. Default is `[]`.
      */
-    public $options = ['role' => 'form'];
+    public $options = [];
     /**
      * @var string the form layout. Either 'default', 'horizontal' or 'inline'.
      * By choosing a layout, an appropriate default field configuration is applied. This will
@@ -88,7 +88,7 @@ class ActiveForm extends \yii\widgets\ActiveForm
 
 
     /**
-     * @inheritdoc
+     * {@inheritdoc}
      */
     public function init()
     {
@@ -103,7 +103,7 @@ class ActiveForm extends \yii\widgets\ActiveForm
     }
 
     /**
-     * @inheritdoc
+     * {@inheritdoc}
      * @return ActiveField the created ActiveField object
      */
     public function field($model, $attribute, $options = [])