+++ src/Collapse.php
@@ -58,17 +58,58 @@ class Collapse extends Widget
      * group with the following structure:
      *
      * - label: string, required, the group header label.
-     * - encode: boolean, optional, whether this label should be HTML-encoded. This param will override
+     * - encode: bool, optional, whether this label should be HTML-encoded. This param will override
      *   global `$this->encodeLabels` param.
      * - content: array|string|object, required, the content (HTML) of the group
      * - options: array, optional, the HTML attributes of the group
      * - contentOptions: optional, the HTML attributes of the group's content
+     *
+     * Since version 2.0.7 you may also specify this property as key-value pairs, where the key refers to the
+     * `label` and the value refers to `content`. If value is a string it is interpreted as label. If it is
+     * an array, it is interpreted as explained above.
+     *
+     * For example:
+     *
+     * ```php
+     * echo Collapse::widget([
+     *     'items' => [
+     *       'Introduction' => 'This is the first collapsable menu',
+     *       'Second panel' => [
+     *           'content' => 'This is the second collapsable menu',
+     *       ],
+     *       [
+     *           'label' => 'Third panel',
+     *           'content' => 'This is the third collapsable menu',
+     *       ],
+     *   ]
+     * ])
+     * ```
      */
     public $items = [];
     /**
-     * @var boolean whether the labels for header items should be HTML-encoded.
+     * @var bool whether the labels for header items should be HTML-encoded.
      */
     public $encodeLabels = true;
+    /**
+     * @var bool whether to close other items if an item is opened. Defaults to `true` which causes an
+     * accordion effect. Set this to `false` to allow keeping multiple items open at once.
+     * @since 2.0.7
+     */
+    public $autoCloseItems = true;
+    /**
+     * @var string the HTML options for the item toggle tag. Key 'tag' might be used here for the tag name specification.
+     * For example:
+     *
+     * ```php
+     * [
+     *     'tag' => 'div',
+     *     'class' => 'custom-toggle',
+     * ]
+     * ```
+     *
+     * @since 2.0.8
+     */
+    public $itemToggleOptions = [];
 
     /**
      * Renders the widget.
@@ -92,9 +133,16 @@ class Collapse extends Widget
     {
         $items = [];
         $index = 0;
-        foreach ($this->items as $item) {
+        foreach ($this->items as $key => $item) {
+            if (!is_array($item)) {
+                $item = ['content' => $item];
+            }
             if (!array_key_exists('label', $item)) {
-                throw new InvalidConfigException("The 'label' option is required.");
+                if (is_int($key)) {
+                    throw new InvalidConfigException("The 'label' option is required.");
+                } else {
+                    $item['label'] = $key;
+                }
             }
             $header = $item['label'];
             $options = ArrayHelper::getValue($item, 'options', []);
@@ -109,7 +157,7 @@ class Collapse extends Widget
      * Renders a single collapsible item group
      * @param string $header a label of the item group [[items]]
      * @param array $item a single item from [[items]]
-     * @param integer $index the item index as each item group content must have an id
+     * @param int $index the item index as each item group content must have an id
      * @return string the rendering result
      * @throws InvalidConfigException
      */
@@ -126,11 +174,23 @@ class Collapse extends Widget
                 $header = Html::encode($header);
             }
 
-            $headerToggle = Html::a($header, '#' . $id, [
-                    'class' => 'collapse-toggle',
-                    'data-toggle' => 'collapse',
-                    'data-parent' => '#' . $this->options['id']
-                ]) . "\n";
+            $itemToggleOptions = array_merge([
+                'tag' => 'a',
+                'data-toggle' => 'collapse',
+            ], $this->itemToggleOptions);
+            Html::addCssClass($itemToggleOptions, ['widget' => 'collapse-toggle']);
+
+            if ($this->autoCloseItems) {
+                $itemToggleOptions['data-parent'] = '#' . $this->options['id'];
+            }
+
+            $itemToggleTag = ArrayHelper::remove($itemToggleOptions, 'tag', 'a');
+            if ($itemToggleTag === 'a') {
+                $headerToggle = Html::a($header, '#' . $id, $itemToggleOptions) . "\n";
+            } else {
+                $itemToggleOptions['data-target'] = '#' . $id;
+                $headerToggle = Html::tag($itemToggleTag, $header, $itemToggleOptions) . "\n";
+            }
 
             $header = Html::tag('h4', $headerToggle, ['class' => 'card-title']);