+++ src/Nav.php
@@ -59,30 +59,30 @@ class Nav extends Widget
      *
      * - label: string, required, the nav item label.
      * - url: optional, the item's URL. Defaults to "#".
-     * - visible: boolean, optional, whether this menu item is visible. Defaults to true.
+     * - visible: bool, optional, whether this menu item is visible. Defaults to true.
      * - linkOptions: array, optional, the HTML attributes of the item's link.
      * - options: array, optional, the HTML attributes of the item container (LI).
-     * - active: boolean, optional, whether the item should be on active state or not.
+     * - active: bool, optional, whether the item should be on active state or not.
      * - dropDownOptions: array, optional, the HTML options that will passed to the [[Dropdown]] widget.
      * - items: array|string, optional, the configuration array for creating a [[Dropdown]] widget,
      *   or a string representing the dropdown menu. Note that Bootstrap does not support sub-dropdown menus.
-     * - encode: boolean, optional, whether the label will be HTML-encoded. If set, supersedes the $encodeLabels option for only this item.
+     * - encode: bool, optional, whether the label will be HTML-encoded. If set, supersedes the $encodeLabels option for only this item.
      *
      * If a menu item is a string, it will be rendered directly without HTML encoding.
      */
     public $items = [];
     /**
-     * @var boolean whether the nav items labels should be HTML-encoded.
+     * @var bool whether the nav items labels should be HTML-encoded.
      */
     public $encodeLabels = true;
     /**
-     * @var boolean whether to automatically activate items according to whether their route setting
+     * @var bool whether to automatically activate items according to whether their route setting
      * matches the currently requested route.
      * @see isItemActive
      */
     public $activateItems = true;
     /**
-     * @var boolean whether to activate parent menu items when one of the corresponding child menu items is active.
+     * @var bool whether to activate parent menu items when one of the corresponding child menu items is active.
      */
     public $activateParents = false;
     /**
@@ -106,10 +106,10 @@ class Nav extends Widget
      */
     public $dropDownCaret;
     /**
-     * @var string name of a class to use for rendering dropdowns withing this widget. Defaults to [[Dropdown]].
-	 * @since 2.0.7
+     * @var string name of a class to use for rendering dropdowns within this widget. Defaults to [[Dropdown]].
+     * @since 2.0.7
      */
-	public $dropdownClass = 'yii\bootstrap\Dropdown';
+    public $dropdownClass = 'yii\bootstrap\Dropdown';
 
 
     /**
@@ -214,9 +214,9 @@ class Nav extends Widget
      */
     protected function renderDropdown($items, $parentItem)
     {
-		/** @var Widget $dropdownClass */
-		$dropdownClass = $this->dropdownClass;
-		return $dropdownClass::widget([
+        /** @var Widget $dropdownClass */
+        $dropdownClass = $this->dropdownClass;
+        return $dropdownClass::widget([
             'options' => ArrayHelper::getValue($parentItem, 'dropDownOptions', []),
             'items' => $items,
             'encodeLabels' => $this->encodeLabels,
@@ -228,15 +228,27 @@ class Nav extends Widget
     /**
      * Check to see if a child item is active optionally activating the parent.
      * @param array $items @see items
-     * @param boolean $active should the parent be active too
+     * @param bool $active should the parent be active too
      * @return array @see items
      */
     protected function isChildActive($items, &$active)
     {
         foreach ($items as $i => $child) {
+            if (is_array($child) && !ArrayHelper::getValue($child, 'visible', true)) {
+                continue;
+            }
             if (ArrayHelper::remove($items[$i], 'active', false) || $this->isItemActive($child)) {
                 Html::addCssClass($items[$i]['options'], 'active');
                 if ($this->activateParents) {
+                    $active = true;
+                }
+            }
+            $childItems = ArrayHelper::getValue($child, 'items');
+            if (is_array($childItems)) {
+                $activeParent = false;
+                $items[$i]['items'] = $this->isChildActive($childItems, $activeParent);
+                if ($activeParent) {
+                    Html::addCssClass($items[$i]['options'], 'active');
                     $active = true;
                 }
             }
@@ -252,7 +264,7 @@ class Nav extends Widget
      * Only when its route and parameters match [[route]] and [[params]], respectively, will a menu item
      * be considered active.
      * @param array $item the menu item to be checked
-     * @return boolean whether the menu item is active
+     * @return bool whether the menu item is active
      */
     protected function isItemActive($item)
     {