+++ src/ToggleButtonGroup.php
@@ -46,13 +46,13 @@ class ToggleButtonGroup extends InputWidget
      */
     public $labelOptions = [];
     /**
-     * @var boolean whether the items labels should be HTML-encoded.
+     * @var bool whether the items labels should be HTML-encoded.
      */
     public $encodeLabels = true;
 
 
     /**
-     * @inheritdoc
+     * {@inheritdoc}
      */
     public function init()
     {
@@ -63,7 +63,7 @@ class ToggleButtonGroup extends InputWidget
     }
 
     /**
-     * @inheritdoc
+     * {@inheritdoc}
      */
     public function run()
     {
@@ -90,10 +90,10 @@ class ToggleButtonGroup extends InputWidget
 
     /**
      * Default callback for checkbox/radio list item rendering.
-     * @param integer $index item index.
+     * @param int $index item index.
      * @param string $label item label.
      * @param string $name input name.
-     * @param boolean $checked whether value is checked or not.
+     * @param bool $checked whether value is checked or not.
      * @param string $value input value.
      * @return string generated HTML.
      * @see Html::checkbox()