+++ tests/ActiveFieldTest.php
@@ -6,6 +6,7 @@ use yii\base\DynamicModel;
 use yii\bootstrap\ActiveField;
 use yii\bootstrap\ActiveForm;
 use Yii;
+use yiiunit\extensions\bootstrap\data\ExtendedActiveField;
 
 class ActiveFieldTest extends TestCase
 {
@@ -36,7 +37,8 @@ class ActiveFieldTest extends TestCase
 
         $this->helperModel = new DynamicModel(['attributeName']);
         ob_start();
-        $this->helperForm = new ActiveForm(['action' => '/something']);
+        $this->helperForm = ActiveForm::begin(['action' => '/something', 'enableClientScript' => false]);
+        ActiveForm::end();
         ob_end_clean();
 
         $this->activeField = new ActiveField(['form' => $this->helperForm]);
@@ -109,4 +111,50 @@ HTML;
 
         $this->assertContains('data-attribute="test"', $content);
     }
-}
+
+    public function testHorizontalCssClasses()
+    {
+        $this->helperForm->layout = 'horizontal';
+
+        $activeField = new ActiveField(['form' => $this->helperForm]);
+        $activeField->model = $this->helperModel;
+        $activeField->attribute = $this->attributeName;
+
+        $html = $activeField->render();
+        $expectedHtml = <<<EXPECTED
+<div class="form-group field-dynamicmodel-attributename">
+<label class="control-label col-sm-3" for="dynamicmodel-attributename">Attribute Name</label>
+<div class="col-sm-6">
+<input type="text" id="dynamicmodel-attributename" class="form-control" name="DynamicModel[attributeName]">
+<p class="help-block help-block-error "></p>
+</div>
+
+</div>
+EXPECTED;
+        $this->assertEqualsWithoutLE($expectedHtml, $html);
+    }
+
+    /**
+     * @depends testHorizontalCssClasses
+     */
+    public function testHorizontalCssClassesOverride()
+    {
+        $this->helperForm->layout = 'horizontal';
+
+        $activeField = new ExtendedActiveField(['form' => $this->helperForm]);
+        $activeField->model = $this->helperModel;
+        $activeField->attribute = $this->attributeName;
+
+        $html = $activeField->render();
+        $expectedHtml = <<<EXPECTED
+<div class="form-group field-dynamicmodel-attributename">
+<label class="control-label col-md-4" for="dynamicmodel-attributename">Attribute Name</label>
+<div class="col-md-6">
+<input type="text" id="dynamicmodel-attributename" class="form-control" name="DynamicModel[attributeName]">
+<p class="help-block help-block-error col-md-3"></p>
+</div>
+
+</div>
+EXPECTED;
+        $this->assertEqualsWithoutLE($expectedHtml, $html);
+    }