+++ tests/bootstrap.php
@@ -12,4 +12,6 @@ require_once(__DIR__ . '/../vendor/autoload.php');
 require_once(__DIR__ . '/../vendor/yiisoft/yii2/Yii.php');
 
 Yii::setAlias('@yiiunit/extensions/bootstrap', __DIR__);
-Yii::setAlias('@yii/bootstrap', dirname(__DIR__));
+Yii::setAlias('@yii/bootstrap', dirname(__DIR__) . '/src');
+
+require_once(__DIR__ . '/compatibility.php');